video_processor/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
aiohttp
requests
python-multipart
# For FFmpeg
ffmpeg-python
